Refactor agent-progress.tsx: Extract 7 nested components into separate files
Extract nested components from agent-progress.tsx into a new agent-progress/ directory to improve code organization and maintainability. This reduces the main file from 2312 lines to 1292 lines (1020 line reduction).

Components extracted:
1. CompactMessage.tsx - Message rendering with TTS, copy, expand/collapse
2. ToolExecutionBubble.tsx - Tool call/result display
3. AssistantWithToolsBubble.tsx - Unified assistant view with tools
4. formatArgumentsPreview.ts - Utility function for argument preview
5. ToolApprovalBubble.tsx - Tool approval/denial UI with keyboard shortcuts
6. RetryStatusBanner.tsx - API retry countdown display
7. StreamingContentBubble.tsx - Streaming response display

All components properly exported via index.ts with TypeScript interfaces.
