+++ apps/desktop/src/renderer/src/components/agent-progress.tsx
@@ -1,21 +1,25 @@
 import React, { useEffect, useRef, useState } from "react"
 import { cn } from "@renderer/lib/utils"
 import { AgentProgressUpdate } from "../../../shared/types"
-import { ChevronDown, ChevronUp, ChevronRight, X, AlertTriangle, Minimize2, Shield, Check, XCircle, Loader2, Clock, Copy, CheckCheck, GripHorizontal, Activity, Moon, Maximize2, RefreshCw, ExternalLink } from "lucide-react"
-import { MarkdownRenderer } from "@renderer/components/markdown-renderer"
+import { ChevronDown, ChevronUp, ChevronRight, X, AlertTriangle, Minimize2, GripHorizontal, Moon, Maximize2, RefreshCw, ExternalLink, Shield, Activity, XCircle, Check } from "lucide-react"
 import { Button } from "./ui/button"
 import { Badge } from "./ui/badge"
 import { tipcClient } from "@renderer/lib/tipc-client"
 import { useAgentStore, useConversationStore, useMessageQueue } from "@renderer/stores"
-import { AudioPlayer } from "@renderer/components/audio-player"
-import { useConfigQuery } from "@renderer/lib/queries"
 import { useTheme } from "@renderer/contexts/theme-context"
 import { logUI, logExpand } from "@renderer/lib/debug"
 import { TileFollowUpInput } from "./tile-follow-up-input"
 import { OverlayFollowUpInput } from "./overlay-follow-up-input"
 import { MessageQueuePanel } from "@renderer/components/message-queue-panel"
 import { useResizable, TILE_DIMENSIONS } from "@renderer/hooks/use-resizable"
-import { getToolResultsSummary } from "@speakmcp/shared"
+import {
+  CompactMessage,
+  ToolExecutionBubble,
+  AssistantWithToolsBubble,
+  ToolApprovalBubble,
+  RetryStatusBanner,
+  StreamingContentBubble
+} from "./agent-progress/index"
 
 interface AgentProgressProps {
   progress: AgentProgressUpdate | null
@@ -77,1027 +81,6 @@ type DisplayItem =
       text: string
       isStreaming: boolean
     } }
-
-
-// Compact message component for space efficiency
-const CompactMessage: React.FC<{
-  message: {
-    role: "user" | "assistant" | "tool"
-    content: string
-    isComplete?: boolean
-    isThinking?: boolean
-    toolCalls?: Array<{ name: string; arguments: any }>
-    toolResults?: Array<{ success: boolean; content: string; error?: string }>
-    timestamp: number
-  }
-  isLast: boolean
-  isComplete: boolean
-  hasErrors: boolean
-  wasStopped?: boolean
-  isExpanded: boolean
-  onToggleExpand: () => void
-  /** Variant controls TTS auto-play - only 'overlay' variant auto-plays TTS to prevent double playback */
-  variant?: "default" | "overlay" | "tile"
-}> = ({ message, isLast, isComplete, hasErrors, wasStopped = false, isExpanded, onToggleExpand, variant = "default" }) => {
-  const [audioData, setAudioData] = useState<ArrayBuffer | null>(null)
-  const [isGeneratingAudio, setIsGeneratingAudio] = useState(false)
-  const [ttsError, setTtsError] = useState<string | null>(null)
-  const [isCopied, setIsCopied] = useState(false)
-  const copyTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const configQuery = useConfigQuery()
-
-  // Cleanup copy timeout on unmount
-  useEffect(() => {
-    return () => {
-      if (copyTimeoutRef.current) {
-        clearTimeout(copyTimeoutRef.current)
-      }
-    }
-  }, [])
-
-  // Copy to clipboard handler
-  const handleCopyResponse = async (e: React.MouseEvent) => {
-    e.stopPropagation()
-    try {
-      await navigator.clipboard.writeText(message.content)
-      setIsCopied(true)
-      // Clear any existing timeout before setting a new one
-      if (copyTimeoutRef.current) {
-        clearTimeout(copyTimeoutRef.current)
-      }
-      copyTimeoutRef.current = setTimeout(() => setIsCopied(false), 2000)
-    } catch (err) {
-      console.error("Failed to copy response:", err)
-    }
-  }
-
-  const displayResults = (message.toolResults || []).filter(
-    (r) =>
-      (r.error && r.error.trim().length > 0) ||
-      (r.content && r.content.trim().length > 0),
-  )
-  const hasExtras =
-    (message.toolCalls?.length ?? 0) > 0 ||
-    displayResults.length > 0
-  const shouldCollapse = (message.content?.length ?? 0) > 100 || hasExtras
-
-  // TTS functionality
-  const generateAudio = async (): Promise<ArrayBuffer> => {
-    if (!configQuery.data?.ttsEnabled) {
-      throw new Error("TTS is not enabled")
-    }
-
-    setIsGeneratingAudio(true)
-    setTtsError(null)
-
-    try {
-      const result = await tipcClient.generateSpeech({
-        text: message.content,
-      })
-      setAudioData(result.audio)
-      return result.audio
-    } catch (error) {
-      console.error("[TTS UI] Failed to generate TTS audio:", error)
-
-      // Set user-friendly error message
-      let errorMessage = "Failed to generate audio"
-      if (error instanceof Error) {
-        if (error.message.includes("API key")) {
-          errorMessage = "TTS API key not configured"
-        } else if (error.message.includes("terms acceptance")) {
-          errorMessage = "Groq TTS model requires terms acceptance. Please visit the Groq console to accept terms for the PlayAI TTS model."
-        } else if (error.message.includes("rate limit")) {
-          errorMessage = "Rate limit exceeded. Please try again later"
-        } else if (error.message.includes("network") || error.message.includes("fetch")) {
-          errorMessage = "Network error. Please check your connection"
-        } else if (error.message.includes("validation")) {
-          errorMessage = "Text content is not suitable for TTS"
-        } else {
-          errorMessage = `TTS error: ${error.message}`
-        }
-      }
-
-      setTtsError(errorMessage)
-      throw error
-    } finally {
-      setIsGeneratingAudio(false)
-    }
-  }
-
-  // Check if TTS should be shown for this message
-  const shouldShowTTS = message.role === "assistant" && isComplete && isLast && configQuery.data?.ttsEnabled
-
-  // Auto-play TTS when assistant message completes (but NOT if agent was stopped by kill switch)
-  //
-  // TTS AUTO-PLAY STRATEGY (fixes #557 - double TTS playback):
-  // - Only auto-play in "overlay" variant (panel window) to prevent double playback
-  // - The panel window is the primary interaction point for agent sessions
-  // - When a non-snoozed session is active, the panel window is automatically shown
-  // - Snoozed sessions don't show the panel, and intentionally don't auto-play TTS
-  //   (they run silently in background - user can unsnooze to see/hear them)
-  // - The "default" variant (main window ConversationDisplay) and "tile" variant (session tiles)
-  //   never auto-play TTS - they are for viewing/managing, not primary interaction
-  useEffect(() => {
-    // Only auto-generate and play TTS in overlay variant to prevent double playback
-    const shouldAutoPlay = variant === "overlay"
-    if (shouldAutoPlay && shouldShowTTS && configQuery.data?.ttsAutoPlay && !audioData && !isGeneratingAudio && !ttsError && !wasStopped) {
-      generateAudio().catch((error) => {
-        // Error is already handled in generateAudio function
-      })
-    }
-  }, [shouldShowTTS, configQuery.data?.ttsAutoPlay, audioData, isGeneratingAudio, ttsError, wasStopped, variant])
-
-  const getRoleStyle = () => {
-    switch (message.role) {
-      case "user":
-        return "border-l-2 border-blue-400 bg-blue-400/5"
-      case "assistant":
-        return isComplete && isLast && !hasErrors
-          ? "border-l-2 border-green-400 bg-green-400/5"
-          : "border-l-2 border-gray-400 bg-gray-400/5"
-      case "tool":
-        return "border-l-2 border-orange-400 bg-orange-400/5"
-    }
-  }
-
-  const getRoleIcon = () => {
-    switch (message.role) {
-      case "user": return "👤"
-      case "assistant": return "🤖"
-      case "tool": return "🔧"
-    }
-  }
-
-  const handleToggleExpand = () => {
-    if (shouldCollapse) {
-      onToggleExpand()
-    }
-  }
-
-  const handleChevronClick = (e: React.MouseEvent) => {
-    e.stopPropagation() // Prevent triggering the message click
-    onToggleExpand()
-  }
-
-  return (
-    <div className={cn(
-      "rounded text-xs transition-all duration-200",
-      getRoleStyle(),
-      !isExpanded && shouldCollapse && "hover:bg-muted/20",
-      shouldCollapse && "cursor-pointer"
-    )}>
-      <div
-        className="flex items-start gap-2 px-2 py-1 text-left"
-        onClick={handleToggleExpand}
-      >
-        <span className="opacity-60 mt-0.5 flex-shrink-0">{getRoleIcon()}</span>
-        <div className="flex-1 min-w-0">
-          <div className={cn(
-            "leading-relaxed text-left",
-            !isExpanded && shouldCollapse && "line-clamp-2"
-          )}>
-          <MarkdownRenderer content={(message.content ?? "").trim()} />
-          </div>
-          {hasExtras && isExpanded && (
-            <div className="mt-2 space-y-2 text-left">
-              {message.toolCalls && message.toolCalls.length > 0 && (
-                <div className="space-y-2">
-                  <div className="text-xs font-semibold opacity-70">Tool Calls ({message.toolCalls.length}):</div>
-                  {message.toolCalls.map((toolCall, index) => (
-                    <div
-                      key={index}
-                      className="rounded-lg border border-border/30 bg-muted/20 p-2 text-xs"
-                    >
-                      <div className="flex items-center justify-between mb-1">
-                        <span className="font-mono font-semibold text-primary">
-                          {toolCall.name}
-                        </span>
-                        <Badge variant="outline" className="text-xs">
-                          Tool {index + 1}
-                        </Badge>
-                      </div>
-                      {toolCall.arguments && (
-                        <div>
-                          <div className="mb-1 text-xs font-medium opacity-70">
-                            Parameters:
-                          </div>
-                          <pre className="rounded bg-muted/50 p-2 overflow-auto text-xs whitespace-pre-wrap max-h-80 scrollbar-thin">
-                            {JSON.stringify(toolCall.arguments, null, 2)}
-                          </pre>
-                        </div>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              )}
-              {displayResults.length > 0 && (
-                <div className="space-y-2">
-                  <div className="text-xs font-semibold opacity-70">Tool Results ({displayResults.length}):</div>
-                  {displayResults.map((result, index) => (
-                    <div
-                      key={index}
-                      className={cn(
-                        "rounded-lg border p-2 text-xs",
-                        result.success
-                          ? "border-green-200/50 bg-green-50/30 text-green-800 dark:border-green-700/50 dark:bg-green-950/40 dark:text-green-200"
-                          : "border-red-200/50 bg-red-50/30 text-red-800 dark:border-red-700/50 dark:bg-red-950/40 dark:text-red-200",
-                      )}
-                    >
-                      <div className="flex items-center justify-between mb-1">
-                        <span className="font-semibold">
-                          {result.success ? "✅ Success" : "❌ Error"}
-                        </span>
-                        <div className="flex items-center gap-2">
-                          <span className="text-[10px] opacity-60 font-mono">
-                            {(result.content?.length || 0).toLocaleString()} chars
-                          </span>
-                          <Badge variant="outline" className="text-xs">
-                            Result {index + 1}
-                          </Badge>
-                        </div>
-                      </div>
-
-                      <div className="space-y-2">
-                        <div>
-                          <div className="text-xs font-medium opacity-70 mb-1">
-                            Content:
-                          </div>
-                          <pre className="rounded bg-muted/30 p-2 overflow-auto text-xs whitespace-pre-wrap break-all max-h-80 scrollbar-thin">
-                            {result.content || "No content returned"}
-                          </pre>
-                        </div>
-
-                        {result.error && (
-                          <div>
-                            <div className="text-xs font-medium text-destructive mb-1">
-                              Error Details:
-                            </div>
-                            <pre className="rounded bg-destructive/10 p-2 overflow-auto text-xs whitespace-pre-wrap break-all max-h-60 scrollbar-thin">
-                              {result.error}
-                            </pre>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </div>
-          )}
-
-          {/* TTS Audio Player - only show for completed assistant messages */}
-          {shouldShowTTS && (
-            <div className="mt-2">
-              <AudioPlayer
-                audioData={audioData || undefined}
-                text={message.content}
-                onGenerateAudio={generateAudio}
-                isGenerating={isGeneratingAudio}
-                error={ttsError}
-                compact={true}
-                autoPlay={configQuery.data?.ttsAutoPlay ?? true}
-              />
-              {ttsError && (
-                <div className="mt-1 rounded-md bg-red-50 p-2 text-xs text-red-700 dark:bg-red-900/20 dark:text-red-300">
-                  <span className="font-medium">Audio generation failed:</span>{" "}
-                  {ttsError.includes("terms acceptance") ? (
-                    <>
-                      Groq TTS model requires terms acceptance.{" "}
-                      <a
-                        href="https://console.groq.com/playground?model=playai-tts"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="underline hover:no-underline"
-                      >
-                        Click here to accept terms
-                      </a>{" "}
-                      for the PlayAI TTS model.
-                    </>
-                  ) : (
-                    ttsError
-                  )}
-                </div>
-              )}
-            </div>
-
-
-          )}
-
-
-        </div>
-        <div className="flex items-center gap-1 flex-shrink-0">
-          {/* Copy button for user prompts and final assistant response */}
-          {(message.role === "user" || (message.role === "assistant" && isLast && isComplete)) && (
-            <button
-              onClick={handleCopyResponse}
-              className="p-1 rounded hover:bg-muted/30 transition-colors"
-              title={isCopied ? "Copied!" : message.role === "user" ? "Copy prompt" : "Copy response"}
-              aria-label={isCopied ? "Copied!" : message.role === "user" ? "Copy prompt" : "Copy response"}
-            >
-              {isCopied ? (
-                <CheckCheck className="h-3 w-3 text-green-500" />
-              ) : (
-                <Copy className="h-3 w-3 opacity-60 hover:opacity-100" />
-              )}
-            </button>
-          )}
-          {shouldCollapse && (
-            <button
-              onClick={handleChevronClick}
-              className="p-1 rounded hover:bg-muted/30 transition-colors"
-            >
-              {isExpanded ? (
-                <ChevronUp className="h-3 w-3" />
-              ) : (
-                <ChevronDown className="h-3 w-3" />
-              )}
-            </button>
-          )}
-        </div>
-      </div>
-    </div>
-  )
-}
-
-// Unified Tool Execution bubble combining call + response
-const ToolExecutionBubble: React.FC<{
-  execution: {
-    timestamp: number
-    calls: Array<{ name: string; arguments: any }>
-    results: Array<{ success: boolean; content: string; error?: string }>
-  }
-  isExpanded: boolean
-  onToggleExpand: () => void
-}> = ({ execution, isExpanded, onToggleExpand }) => {
-  const [showInputs, setShowInputs] = useState(false)
-  const [showOutputs, setShowOutputs] = useState(false)
-
-  // Collapsed by default; expand to show details
-  useEffect(() => {
-    if (isExpanded) {
-      setShowInputs(true)
-      setShowOutputs(true)
-    } else {
-      setShowInputs(false)
-      setShowOutputs(false)
-    }
-  }, [isExpanded, execution])
-
-  const isPending = execution.results.length === 0
-  const allSuccess = execution.results.length > 0 && execution.results.every((r) => r.success)
-  const hasErrors = execution.results.length > 0 && execution.results.some((r) => !r.success)
-  const headerTitle = execution.calls.map((c) => c.name).join(", ") || "Tool Execution"
-
-  const copy = async (text: string) => {
-    try {
-      await navigator.clipboard?.writeText(text)
-    } catch {}
-  }
-
-  const handleToggleExpand = () => onToggleExpand()
-  const handleChevronClick = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    onToggleExpand()
-  }
-
-  // Handle hide/show buttons with event propagation stopped
-  const handleToggleInputs = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setShowInputs((v) => !v)
-  }
-
-  const handleToggleOutputs = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setShowOutputs((v) => !v)
-  }
-
-  const handleCopy = (e: React.MouseEvent, text: string) => {
-    e.stopPropagation()
-    copy(text)
-  }
-
-
-  // Generate preview for collapsed state - prioritize showing results when available
-  const collapsedInputPreview = (() => {
-    if (isExpanded) return null
-    // Create a summary of the first tool call's parameters
-    const firstCall = execution.calls[0]
-    if (!firstCall?.arguments) return null
-    return formatArgumentsPreview(firstCall.arguments)
-  })()
-
-  // Generate result summary for collapsed state
-  const collapsedResultSummary = (() => {
-    if (isExpanded || isPending) return null
-    if (execution.results.length === 0) return null
-    // Convert to the expected ToolResult format
-    const toolResults = execution.results.map(r => ({
-      success: r.success,
-      content: r.content,
-      error: r.error,
-    }))
-    return getToolResultsSummary(toolResults)
-  })()
-
-  return (
-    <div
-      className={cn(
-        "rounded-lg border p-2 text-xs",
-        isPending
-          ? "border-blue-200/50 bg-blue-50/30 text-blue-800 dark:border-blue-700/50 dark:bg-blue-950/40 dark:text-blue-200"
-          : allSuccess
-            ? "border-green-200/50 bg-green-50/30 text-green-800 dark:border-green-700/50 dark:bg-green-950/40 dark:text-green-200"
-            : "border-red-200/50 bg-red-50/30 text-red-800 dark:border-red-700/50 dark:bg-red-950/40 dark:text-red-200",
-      )}
-    >
-      <div
-        className="mb-1 flex items-center justify-between px-1 py-1 cursor-pointer hover:bg-muted/20 rounded"
-        onClick={handleToggleExpand}
-        aria-expanded={isExpanded}
-      >
-        <div className="flex items-center gap-2 min-w-0 flex-1">
-          <span className="font-mono font-semibold truncate">{headerTitle}</span>
-          {!isExpanded && (
-            <Badge variant="outline" className="text-[10px] flex-shrink-0">
-              {isPending ? "Pending..." : allSuccess ? "✓" : "✗"}
-            </Badge>
-          )}
-          {isExpanded && (
-            <Badge variant="outline" className="text-[10px]">
-              {isPending ? "Pending..." : allSuccess ? "Success" : "With errors"}
-            </Badge>
-          )}
-        </div>
-        <div className="flex items-center gap-2 flex-shrink-0">
-          {isExpanded && (
-            <span className="opacity-60 text-[10px]">{new Date(execution.timestamp).toLocaleTimeString()}</span>
-          )}
-          <button
-            onClick={handleChevronClick}
-            className="p-1 rounded hover:bg-muted/30 transition-colors"
-            aria-label={isExpanded ? "Collapse" : "Expand"}
-          >
-            {isExpanded ? <ChevronDown className="h-3 w-3" /> : <ChevronRight className="h-3 w-3" />}
-          </button>
-        </div>
-      </div>
-
-      {/* Collapsed preview - show result summary when available, otherwise show input parameters */}
-      {!isExpanded && (collapsedResultSummary || collapsedInputPreview) && (
-        <div className="px-1 pb-1 text-[10px] opacity-80 truncate" title={collapsedResultSummary || collapsedInputPreview || ''}>
-          {collapsedResultSummary ? (
-            <span className="font-medium">{collapsedResultSummary}</span>
-          ) : (
-            <span className="font-mono opacity-70">{collapsedInputPreview}</span>
-          )}
-        </div>
-      )}
-
-      {isExpanded && (
-        <>
-          {/* Inputs */}
-          <div className="rounded-md bg-blue-50/40 dark:bg-blue-900/10 border border-blue-200/40 dark:border-blue-800/40 p-2 mb-2">
-            <div className="flex items-center justify-between">
-              <div className="text-[11px] font-semibold opacity-80">Call Parameters</div>
-              <div className="flex items-center gap-1">
-                <Button size="sm" variant="ghost" className="h-6 px-2" onClick={handleToggleInputs}>
-                  {showInputs ? "Hide" : "Show"}
-                </Button>
-                <Button size="sm" variant="ghost" className="h-6 px-2" onClick={(e) => handleCopy(e, JSON.stringify(execution.calls, null, 2))}>
-                  Copy
-                </Button>
-              </div>
-            </div>
-            {showInputs && (
-              <div className="mt-1 space-y-2">
-                {execution.calls.map((c, idx) => (
-                  <div key={idx} className="rounded bg-muted/50 p-2 overflow-auto whitespace-pre-wrap max-h-80 scrollbar-thin">
-                    <div className="mb-1 text-[11px] font-medium opacity-70">{c.name}</div>
-                    <pre>{JSON.stringify(c.arguments ?? {}, null, 2)}</pre>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-
-          {/* Outputs */}
-          <div
-            className="rounded-md border p-2"
-            style={{
-              borderColor: isPending ? "rgb(191 219 254 / 0.5)" : allSuccess ? "rgb(187 247 208 / 0.5)" : "rgb(254 202 202 / 0.5)",
-              backgroundColor: isPending ? "rgb(239 246 255 / 0.3)" : allSuccess ? "rgb(240 253 244 / 0.3)" : "rgb(254 242 242 / 0.3)",
-            } as React.CSSProperties}
-          >
-            <div className="flex items-center justify-between">
-              <div className="text-[11px] font-semibold opacity-80">Response</div>
-              {!isPending && (
-                <div className="flex items-center gap-1">
-                  <Button size="sm" variant="ghost" className="h-6 px-2" onClick={handleToggleOutputs}>
-                    {showOutputs ? "Hide" : "Show"}
-                  </Button>
-                  <Button size="sm" variant="ghost" className="h-6 px-2" onClick={(e) => handleCopy(e, JSON.stringify(execution.results, null, 2))}>
-                    Copy
-                  </Button>
-                </div>
-              )}
-            </div>
-            {isPending ? (
-              <div className="mt-2 text-center py-2 text-[11px] opacity-60 italic">
-                Waiting for response...
-              </div>
-            ) : showOutputs && (
-              <div className="mt-1 space-y-2">
-                {execution.results.map((r, idx) => (
-                  <div
-                    key={idx}
-                    className={cn(
-                      "rounded border p-2 text-xs",
-                      r.success
-                        ? "border-green-200/50 bg-green-50/30 dark:border-green-700/50 dark:bg-green-950/30"
-                        : "border-red-200/50 bg-red-50/30 dark:border-red-700/50 dark:bg-red-950/30",
-                    )}
-                  >
-                    <div className="mb-1 flex items-center justify-between">
-                      <span className="font-semibold">{r.success ? "✅ Success" : "❌ Error"}</span>
-                      <div className="flex items-center gap-2">
-                        <span className="text-[10px] opacity-60 font-mono">
-                          {(r.content?.length || 0).toLocaleString()} chars
-                        </span>
-                        <Badge variant="outline" className="text-[10px]">{`Result ${idx + 1}`}</Badge>
-                      </div>
-                    </div>
-                    <div className="space-y-2">
-                      <div>
-                        <div className="text-[11px] font-medium opacity-70 mb-1">Content:</div>
-                        <pre className="rounded bg-muted/30 p-2 overflow-auto whitespace-pre-wrap break-all max-h-80 scrollbar-thin">
-                          {r.content || "No content returned"}
-                        </pre>
-                      </div>
-                      {r.error && (
-                        <div>
-                          <div className="text-[11px] font-medium text-destructive mb-1">Error Details:</div>
-                          <pre className="rounded bg-destructive/10 p-2 overflow-auto whitespace-pre-wrap break-all max-h-60 scrollbar-thin">
-                            {r.error}
-                          </pre>
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        </>
-      )}
-
-
-    </div>
-  )
-}
-
-// Unified Assistant + Tool Execution component - combines thought and tool call as one message
-const AssistantWithToolsBubble: React.FC<{
-  data: {
-    thought: string
-    timestamp: number
-    isComplete: boolean
-    calls: Array<{ name: string; arguments: any }>
-    results: Array<{ success: boolean; content: string; error?: string }>
-  }
-  isExpanded: boolean
-  onToggleExpand: () => void
-}> = ({ data, isExpanded, onToggleExpand }) => {
-  const [showToolDetails, setShowToolDetails] = useState(false)
-
-  const isPending = data.results.length === 0
-  const allSuccess = data.results.length > 0 && data.results.every(r => r.success)
-  const hasThought = data.thought && data.thought.trim().length > 0
-  const shouldCollapse = (data.thought?.length ?? 0) > 100 || data.calls.length > 0
-
-  // Generate result summary for collapsed state
-  const collapsedResultSummary = (() => {
-    if (isExpanded || isPending) return null
-    if (data.results.length === 0) return null
-    const toolResults = data.results.map(r => ({
-      success: r.success,
-      content: r.content,
-      error: r.error,
-    }))
-    return getToolResultsSummary(toolResults)
-  })()
-
-  const handleToggleExpand = () => {
-    if (shouldCollapse) {
-      onToggleExpand()
-    }
-  }
-
-  const handleChevronClick = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    onToggleExpand()
-  }
-
-  const handleToggleToolDetails = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setShowToolDetails(!showToolDetails)
-  }
-
-  // Tool names for display
-  const toolNames = data.calls.map(c => c.name).join(', ')
-  const toolCount = data.calls.length
-
-  return (
-    <div className={cn(
-      "rounded text-xs transition-all duration-200",
-      "border-l-2 border-gray-400 bg-gray-400/5",
-      !isExpanded && shouldCollapse && "hover:bg-muted/20",
-      shouldCollapse && "cursor-pointer"
-    )}>
-      {/* Thought content section */}
-      <div
-        className="flex items-start gap-2 px-2 py-1 text-left"
-        onClick={handleToggleExpand}
-      >
-        <span className="opacity-60 mt-0.5 flex-shrink-0">🤖</span>
-        <div className="flex-1 min-w-0">
-          {hasThought && (
-            <div className={cn(
-              "leading-relaxed text-left",
-              !isExpanded && shouldCollapse && "line-clamp-2"
-            )}>
-              <MarkdownRenderer content={data.thought.trim()} />
-            </div>
-          )}
-
-          {/* Tool execution section - always visible but collapsible */}
-          <div className={cn(
-            "mt-2 rounded-lg border p-2",
-            isPending
-              ? "border-blue-200/50 bg-blue-50/30 text-blue-800 dark:border-blue-700/50 dark:bg-blue-950/40 dark:text-blue-200"
-              : allSuccess
-                ? "border-green-200/50 bg-green-50/30 text-green-800 dark:border-green-700/50 dark:bg-green-950/40 dark:text-green-200"
-                : "border-red-200/50 bg-red-50/30 text-red-800 dark:border-red-700/50 dark:bg-red-950/40 dark:text-red-200",
-          )}>
-            <div
-              className="flex items-center justify-between cursor-pointer"
-              onClick={handleToggleToolDetails}
-            >
-              <div className="flex items-center gap-2 min-w-0 flex-1">
-                <span className="opacity-60">🔧</span>
-                <span className="font-mono font-semibold truncate">
-                  {toolCount === 1 ? data.calls[0].name : `${toolCount} tool calls`}
-                </span>
-                <Badge variant="outline" className="text-[10px] flex-shrink-0">
-                  {isPending ? "Running..." : allSuccess ? "✓" : "✗"}
-                </Badge>
-              </div>
-              <button
-                onClick={handleChevronClick}
-                className="p-1 rounded hover:bg-muted/30 transition-colors flex-shrink-0"
-                aria-label={showToolDetails ? "Collapse" : "Expand"}
-              >
-                {showToolDetails ? <ChevronDown className="h-3 w-3" /> : <ChevronRight className="h-3 w-3" />}
-              </button>
-            </div>
-
-            {/* Collapsed preview - show result summary */}
-            {!showToolDetails && collapsedResultSummary && (
-              <div className="mt-1 text-[10px] opacity-80 truncate">
-                <span className="font-medium">{collapsedResultSummary}</span>
-              </div>
-            )}
-
-            {/* Expanded tool details */}
-            {showToolDetails && (
-              <div className="mt-2 space-y-2">
-                {/* Tool calls */}
-                {data.calls.map((call, idx) => (
-                  <div key={idx} className="rounded bg-muted/30 p-2">
-                    <div className="flex items-center justify-between mb-1">
-                      <span className="font-mono font-semibold text-primary">{call.name}</span>
-                      <Badge variant="outline" className="text-[10px]">Call {idx + 1}</Badge>
-                    </div>
-                    {call.arguments && (
-                      <pre className="rounded bg-muted/50 p-2 overflow-auto text-xs whitespace-pre-wrap max-h-40 scrollbar-thin">
-                        {JSON.stringify(call.arguments, null, 2)}
-                      </pre>
-                    )}
-                  </div>
-                ))}
-
-                {/* Tool results */}
-                {data.results.length > 0 && (
-                  <div className="space-y-2">
-                    <div className="text-[11px] font-semibold opacity-70">Results:</div>
-                    {data.results.map((result, idx) => (
-                      <div
-                        key={idx}
-                        className={cn(
-                          "rounded border p-2",
-                          result.success
-                            ? "border-green-200/50 bg-green-50/20 dark:border-green-700/50 dark:bg-green-950/20"
-                            : "border-red-200/50 bg-red-50/20 dark:border-red-700/50 dark:bg-red-950/20",
-                        )}
-                      >
-                        <div className="flex items-center justify-between mb-1">
-                          <span className="font-semibold">{result.success ? "✅ Success" : "❌ Error"}</span>
-                          <span className="text-[10px] opacity-60 font-mono">
-                            {(result.content?.length || 0).toLocaleString()} chars
-                          </span>
-                        </div>
-                        <pre className="rounded bg-muted/30 p-2 overflow-auto whitespace-pre-wrap break-all max-h-40 scrollbar-thin">
-                          {result.content || "No content returned"}
-                        </pre>
-                        {result.error && (
-                          <div className="mt-2">
-                            <div className="text-[11px] font-medium text-destructive mb-1">Error:</div>
-                            <pre className="rounded bg-destructive/10 p-2 overflow-auto whitespace-pre-wrap break-all max-h-40 scrollbar-thin">
-                              {result.error}
-                            </pre>
-                          </div>
-                        )}
-                      </div>
-                    ))}
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-// Helper function to format tool arguments for preview
-const formatArgumentsPreview = (args: any): string => {
-  if (!args || typeof args !== 'object') return ''
-  const entries = Object.entries(args)
-  if (entries.length === 0) return ''
-
-  // Take first 3 key parameters
-  const preview = entries.slice(0, 3).map(([key, value]) => {
-    let displayValue: string
-    if (typeof value === 'string') {
-      displayValue = value.length > 30 ? value.slice(0, 30) + '...' : value
-    } else if (typeof value === 'object') {
-      displayValue = Array.isArray(value) ? `[${value.length} items]` : '{...}'
-    } else {
-      displayValue = String(value)
-    }
-    return `${key}: ${displayValue}`
-  }).join(', ')
-
-  if (entries.length > 3) {
-    return preview + ` (+${entries.length - 3} more)`
-  }
-  return preview
-}
-
-// Inline Tool Approval bubble - appears in the conversation flow
-const ToolApprovalBubble: React.FC<{
-  approval: {
-    approvalId: string
-    toolName: string
-    arguments: any
-  }
-  onApprove: () => void
-  onDeny: () => void
-  isResponding: boolean
-}> = ({ approval, onApprove, onDeny, isResponding }) => {
-  const [showArgs, setShowArgs] = useState(false)
-  const containerRef = useRef<HTMLDivElement>(null)
-
-  // Keyboard shortcut handler for tool approval
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      // Don't handle if already responding or if user is typing in an input
-      if (isResponding) return
-      const target = e.target as HTMLElement
-      // Ignore when focus is on interactive elements to preserve standard keyboard navigation
-      if (
-        target.tagName === 'INPUT' ||
-        target.tagName === 'TEXTAREA' ||
-        target.tagName === 'BUTTON' ||
-        target.tagName === 'A' ||
-        target.isContentEditable
-      ) {
-        return
-      }
-
-      // Use e.code for more consistent Space detection across browsers/platforms
-      // Space to approve (without modifiers)
-      if (e.code === 'Space' && !e.shiftKey && !e.ctrlKey && !e.altKey && !e.metaKey) {
-        e.preventDefault()
-        onApprove()
-      }
-      // Shift+Space to deny
-      else if (e.code === 'Space' && e.shiftKey && !e.ctrlKey && !e.altKey && !e.metaKey) {
-        e.preventDefault()
-        onDeny()
-      }
-    }
-
-    document.addEventListener('keydown', handleKeyDown)
-    return () => document.removeEventListener('keydown', handleKeyDown)
-  }, [isResponding, onApprove, onDeny])
-
-  // Generate preview text for collapsed view hint
-  const argsPreview = formatArgumentsPreview(approval.arguments)
-
-  return (
-    <div className="rounded-lg border border-amber-300 dark:border-amber-700 bg-amber-50/50 dark:bg-amber-950/30 overflow-hidden">
-      {/* Header */}
-      <div className="flex items-center gap-2 px-3 py-2 bg-amber-100/50 dark:bg-amber-900/30 border-b border-amber-200 dark:border-amber-800">
-        <Shield className="h-3.5 w-3.5 text-amber-600 dark:text-amber-400" />
-        <span className="text-xs font-medium text-amber-800 dark:text-amber-200">
-          {isResponding ? "Processing..." : "Tool Approval Required"}
-        </span>
-        {isResponding && (
-          <Loader2 className="h-3 w-3 text-amber-600 dark:text-amber-400 animate-spin ml-auto" />
-        )}
-      </div>
-
-      {/* Content */}
-      <div ref={containerRef} className={cn("px-3 py-2", isResponding && "opacity-60")}>
-        <div className="flex items-center gap-2 mb-2">
-          <span className="text-xs text-amber-700 dark:text-amber-300">Tool:</span>
-          <code className="text-xs font-mono font-medium text-amber-900 dark:text-amber-100 bg-amber-100 dark:bg-amber-900/50 px-1.5 py-0.5 rounded">
-            {approval.toolName}
-          </code>
-        </div>
-
-        {/* Arguments preview - always visible */}
-        {argsPreview && (
-          <div className="mb-2 text-xs text-amber-700/80 dark:text-amber-300/80 font-mono truncate" title={argsPreview}>
-            {argsPreview}
-          </div>
-        )}
-
-        {/* Expandable arguments */}
-        <div className="mb-3">
-          <button
-            onClick={() => setShowArgs(!showArgs)}
-            className="flex items-center gap-1 text-xs text-amber-600 dark:text-amber-400 hover:text-amber-800 dark:hover:text-amber-200"
-            disabled={isResponding}
-          >
-            <ChevronRight className={cn("h-3 w-3 transition-transform", showArgs && "rotate-90")} />
-            {showArgs ? "Hide" : "View"} full arguments
-          </button>
-          {showArgs && (
-            <pre className="mt-1.5 p-2 text-xs bg-amber-100/70 dark:bg-amber-900/40 rounded overflow-x-auto max-h-32 text-amber-900 dark:text-amber-100">
-              {JSON.stringify(approval.arguments, null, 2)}
-            </pre>
-          )}
-        </div>
-
-        {/* Action buttons with hotkey hints */}
-        <div className="flex items-center gap-2">
-          <Button
-            variant="outline"
-            size="sm"
-            className="h-7 text-xs border-red-300 text-red-700 hover:bg-red-50 dark:border-red-700 dark:text-red-400 dark:hover:bg-red-950/30"
-            onClick={onDeny}
-            disabled={isResponding}
-            title="Press Shift+Space to deny"
-          >
-            <XCircle className="h-3 w-3 mr-1" />
-            Deny
-            <kbd className="ml-1.5 px-1 py-0.5 text-[9px] font-mono bg-red-100 dark:bg-red-900/50 rounded">Shift+Space</kbd>
-          </Button>
-          <Button
-            size="sm"
-            className={cn(
-              "h-7 text-xs text-white",
-              isResponding
-                ? "bg-green-500 cursor-not-allowed"
-                : "bg-green-600 hover:bg-green-700"
-            )}
-            onClick={onApprove}
-            disabled={isResponding}
-            title="Press Space to approve"
-          >
-            {isResponding ? (
-              <>
-                <Loader2 className="h-3 w-3 mr-1 animate-spin" />
-                Processing...
-              </>
-            ) : (
-              <>
-                <Check className="h-3 w-3 mr-1" />
-                Approve
-                <kbd className="ml-1.5 px-1 py-0.5 text-[9px] font-mono bg-green-700 rounded">Space</kbd>
-              </>
-            )}
-          </Button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-// Retry Status Banner - shows when LLM API is being retried (rate limits, network errors)
-const RetryStatusBanner: React.FC<{
-  retryInfo: {
-    isRetrying: boolean
-    attempt: number
-    maxAttempts?: number
-    delaySeconds: number
-    reason: string
-    startedAt: number
-  }
-}> = ({ retryInfo }) => {
-  const [countdown, setCountdown] = useState(retryInfo.delaySeconds)
-
-  // Update countdown timer
-  useEffect(() => {
-    if (!retryInfo.isRetrying) {
-      setCountdown(0)
-      return undefined
-    }
-
-    // Calculate remaining time based on startedAt
-    const updateCountdown = () => {
-      const elapsed = Math.floor((Date.now() - retryInfo.startedAt) / 1000)
-      const remaining = Math.max(0, retryInfo.delaySeconds - elapsed)
-      setCountdown(remaining)
-    }
-
-    updateCountdown()
-    const interval = setInterval(updateCountdown, 1000)
-    return () => clearInterval(interval)
-  }, [retryInfo.isRetrying, retryInfo.startedAt, retryInfo.delaySeconds])
-
-  if (!retryInfo.isRetrying) return null
-
-  const attemptText = retryInfo.maxAttempts
-    ? `Attempt ${retryInfo.attempt}/${retryInfo.maxAttempts}`
-    : `Attempt ${retryInfo.attempt}`
-
-  return (
-    <div className="rounded-lg border border-amber-300 dark:border-amber-700 bg-amber-50/50 dark:bg-amber-950/30 overflow-hidden">
-      {/* Header */}
-      <div className="flex items-center gap-2 px-3 py-2 bg-amber-100/50 dark:bg-amber-900/30 border-b border-amber-200 dark:border-amber-800">
-        <Clock className="h-3.5 w-3.5 text-amber-600 dark:text-amber-400" />
-        <span className="text-xs font-medium text-amber-800 dark:text-amber-200">
-          {retryInfo.reason}
-        </span>
-        <Loader2 className="h-3 w-3 text-amber-600 dark:text-amber-400 animate-spin ml-auto" />
-      </div>
-
-      {/* Content */}
-      <div className="px-3 py-2">
-        <div className="flex items-center justify-between">
-          <span className="text-xs text-amber-700 dark:text-amber-300">
-            {attemptText}
-          </span>
-          <span className="text-xs font-mono font-medium text-amber-900 dark:text-amber-100 bg-amber-100 dark:bg-amber-900/50 px-2 py-0.5 rounded">
-            Retrying in {countdown}s
-          </span>
-        </div>
-        <p className="text-xs text-amber-600 dark:text-amber-400 mt-1.5">
-          The agent will automatically retry when the API is available.
-        </p>
-      </div>
-    </div>
-  )
-}
-
-// Streaming Content Bubble - shows real-time LLM response as it's being generated
-const StreamingContentBubble: React.FC<{
-  streamingContent: {
-    text: string
-    isStreaming: boolean
-  }
-}> = ({ streamingContent }) => {
-  if (!streamingContent.text) return null
-
-  return (
-    <div className="rounded-lg border border-blue-300 dark:border-blue-700 bg-blue-50/50 dark:bg-blue-950/30 overflow-hidden">
-      {/* Header */}
-      <div className="flex items-center gap-2 px-3 py-2 bg-blue-100/50 dark:bg-blue-900/30 border-b border-blue-200 dark:border-blue-800">
-        <Activity className="h-3.5 w-3.5 text-blue-600 dark:text-blue-400" />
-        <span className="text-xs font-medium text-blue-800 dark:text-blue-200">
-          {streamingContent.isStreaming ? "Generating response..." : "Response"}
-        </span>
-        {streamingContent.isStreaming && (
-          <Loader2 className="h-3 w-3 text-blue-600 dark:text-blue-400 animate-spin ml-auto" />
-        )}
-      </div>
-
-      {/* Content */}
-      <div className="px-3 py-2">
-        <div className="text-xs text-blue-900 dark:text-blue-100 whitespace-pre-wrap break-words">
-          <MarkdownRenderer content={streamingContent.text} />
-          {streamingContent.isStreaming && (
-            <span className="inline-block w-1.5 h-3.5 bg-blue-600 dark:bg-blue-400 animate-pulse ml-0.5 align-text-bottom" />
-          )}
-        </div>
-      </div>
-    </div>
-  )
-}
-
 
 export const AgentProgress: React.FC<AgentProgressProps> = ({
   progress,
